allow to search deleted workload logs
